# Introduce multiSelect for ScalarQuantizer