--- a/lucene/core/src/java/org/apache/lucene/util/IntroSelector.java
+++ b/lucene/core/src/java/org/apache/lucene/util/IntroSelector.java
@@ -154,7 +154,8 @@ public abstract class IntroSelector extends Selector {
 
   // Visible for testing.
   void multiSelect(int from, int to, int[] k, int kFrom, int kTo, int maxDepth) {
-    // If there is only 1 k value to select in this group, then use the single-k select method, which does not do recursion
+    // If there is only 1 k value to select in this group, then use the single-k select method,
+    // which does not do recursion
     if (kTo - kFrom == 1) {
       select(from, to, k[kFrom], maxDepth);
       return;
@@ -240,7 +241,7 @@ public abstract class IntroSelector extends Selector {
       // Select the K values contained in the bottom and top partitions.
       int topKFrom = kTo;
       int bottomKTo = kFrom;
-      for (int ki = kTo-1; ki >= kFrom; ki--) {
+      for (int ki = kTo - 1; ki >= kFrom; ki--) {
         if (k[ki] >= i) {
           topKFrom = ki;
         }
@@ -249,11 +250,13 @@ public abstract class IntroSelector extends Selector {
           break;
         }
       }
-      // Recursively select the relevant k-values from the bottom group, if there are any k-values to select there
+      // Recursively select the relevant k-values from the bottom group, if there are any k-values
+      // to select there
       if (bottomKTo > kFrom) {
         multiSelect(from, j + 1, k, kFrom, bottomKTo, maxDepth);
       }
-      // Recursively select the relevant k-values from the top group, if there are any k-values to select there
+      // Recursively select the relevant k-values from the top group, if there are any k-values to
+      // select there
       if (topKFrom < kTo) {
         multiSelect(i, to, k, topKFrom, kTo, maxDepth);
       }