--- a/lucene/core/src/java/org/apache/lucene/util/Selector.java
+++ b/lucene/core/src/java/org/apache/lucene/util/Selector.java
@@ -16,6 +16,8 @@
  */
 package org.apache.lucene.util;
 
+import java.util.Arrays;
+
 /**
  * An implementation of a selection algorithm, ie. computing the k-th greatest value from a
  * collection.
@@ -30,12 +32,35 @@ public abstract class Selector {
    */
   public abstract void select(int from, int to, int k);
 
+  /**
+   * Reorder elements so that the elements at all positions in {@code k} are the same as if all elements were
+   * sorted and all other elements are partitioned around it: {@code [from, k[n])} only contains
+   * elements that are less than or equal to {@code k[n]} and {@code (k[n], to)} only contains elements
+   * that are greater than or equal to {@code k[n]}.
+   */
+  public void select(int from, int to, int[] k) {
+    select(from, to, k[0]);
+  }
+
   void checkArgs(int from, int to, int k) {
     if (k < from) {
       throw new IllegalArgumentException("k must be >= from");
     }
     if (k >= to) {
       throw new IllegalArgumentException("k must be < to");
+    }
+  }
+
+  void checkArgs(int from, int to, int[] k) {
+    if (k.length < 1) {
+      throw new IllegalArgumentException("There must be at least one k to select, none given");
+    }
+    Arrays.sort(k);
+    if (k[0] < from) {
+      throw new IllegalArgumentException("All k must be >= from");
+    }
+    if (k[k.length - 1] >= to) {
+      throw new IllegalArgumentException("All k must be < to");
     }
   }