--- a/lucene/core/src/java/org/apache/lucene/util/Selector.java
+++ b/lucene/core/src/java/org/apache/lucene/util/Selector.java
@@ -16,6 +16,8 @@
  */
 package org.apache.lucene.util;
 
+import java.util.Arrays;
+
 /**
  * An implementation of a selection algorithm, ie. computing the k-th greatest value from a
  * collection.
@@ -39,6 +41,7 @@ public abstract class Selector {
   public void multiSelect(int from, int to, int[] k) {
     // k needs to be sorted, so copy the array
     k = ArrayUtil.copyArray(k);
+    Arrays.sort(k);
     checkMultiArgs(from, to, k);
     multiSelect(from, to, k, 0, k.length);
   }