@HoustonPutman 2024-10-15
So after fixing the innocuous bugs in the implementations, it looks like there is no speed up here. The confidence interval finding can be up to 30% faster or so, but that's such a small portion of the quantization cost, that it really doesn't make a difference at all. (I think the bug was creating bad quantization that made the `findNearestNeighbors` search much faster)

Oh well, good to test out and see at least. Probably no reason to merge if there isn't a Lucene use case that would really benefit from this.